Updated gitnore
We no longer ignore the build folder. We want people to use the ES5 version of this library without the need to lean and use Gulp for transpiling. So we let Travis CI do the transpiling on each commit to the master branch.
It's not the perfect solution but I couldn't find a better one.